Server: Hack in a queue for get-requests
It very much looks like the editor expects to send all it's requests
before it reads any replies. Supporting that implies having an internal
queue for get-requests.

Currently hacked by sending io.Reader's over a channel. It will cause a
lot of open file handles to be around, but that can be fixed if it
proves successful.